automation/requirements.txt
# SignLanguageTranslate Automation Dependencies

# Claude API
anthropic>=0.40.0

# Configuration
pyyaml>=6.0.1

# CLI and Console
rich>=13.7.0
click>=8.1.7

# Templating
jinja2>=3.1.2

# Async Operations
aiosqlite>=0.19.0
aiofiles>=23.2.1
httpx>=0.26.0

# Utilities
python-dateutil>=2.8.2
watchdog>=4.0.0
pillow>=10.2.0